data_modeling_sql/requirements.txt
openai
python-dotenv
requests
boto3
trino